requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
docstring_parser==0.16
fhaviary==0.6.0
filelock==3.16.0
frozenlist==1.4.1
fsspec==2024.9.0
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
importlib_metadata==8.5.0
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
latexcodec==3.0.0
litellm==1.44.24
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.1.0
numpy==2.1.1
openai==1.44.1
packaging==24.1
pybtex==0.24.0
pycryptodome==3.20.0
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
Pygments==2.18.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
setuptools==74.1.2
six==1.16.0
sniffio==1.3.1
tantivy==0.22.0
tenacity==9.0.0
tiktoken==0.7.0
tokenizers==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
wheel==0.44.0
yarl==1.11.1
zipp==3.20.1
